Move logger.ts from client/ -> lib/
