+++ packages/api/src/client/activityApi.ts
@@ -1,6 +1,7 @@
 import { da } from '@urbit/aura';
 import { backOff } from 'exponential-backoff';
 
+import { createDevLogger, runIfDev } from '../lib/logger';
 import { normalizeUrbitColor } from '../lib/utils';
 import type * as db from '../types/models';
 import { BASE_UNREADS_SINGLETON_KEY } from '../types/models';
@@ -13,7 +14,6 @@ import {
   parseGroupId,
   udToDate,
 } from './apiUtils';
-import { createDevLogger, runIfDev } from './logger';
 import { poke, scry, subscribe } from './urbit';
 
 const logger = createDevLogger('activityApi', false);
