+++ packages/api/src/client/harkApi.ts
@@ -1,5 +1,5 @@
+import { createDevLogger } from '../lib/logger';
 import { getCanonicalPostId } from './apiUtils';
-import { createDevLogger } from './logger';
 
 const logger = createDevLogger('harkApi', true);
