+++ packages/api/src/http-api/Urbit.ts
@@ -2,9 +2,9 @@ import { parse, render } from '@urbit/aura';
 import { Atom, Cell, Noun, dejs, enjs, jam } from '@urbit/nockjs';
 
 import { TimeoutError } from '../client';
-import { createDevLogger } from '../client/logger';
 import { EventEmitter } from '../lib/EventEmitter';
 import { readArrayBufferFromBlob } from '../lib/blob';
+import { createDevLogger } from '../lib/logger';
 import { createTimeoutSignal } from '../lib/timeoutSignal';
 import { desig } from '../urbit';
 import { UrbitHttpApiEvent, UrbitHttpApiEventType } from './events';
