+++ packages/api/src/lib/branch.ts
@@ -1,7 +1,7 @@
 import { valid } from '@urbit/aura';
 
 import { getPostInfoFromWer } from '../client/harkApi';
-import { createDevLogger } from '../client/logger';
+import { createDevLogger } from '../lib/logger';
 import { getConstants } from '../types/constants';
 import { preSig } from '../urbit';
 import { normalizeUrbitColor } from './utils';
